use the provided GITHUB_TOKEN instead of a provided one (#14173)
Recently updating these shopify provided github tokens to GITHUB_TOKEN,
which is provided by github in the run context of the job, has been
needed to get old jobs running: I think a credential changed or expired
at the org scope.

Unsure if this will fix the problem, but should give us next steps.

Also added a `workflow_dispatch` to this action with a configurable
branch name, so we can test the deploy flow from PRs in the future, no
more CI golf.